Add pretraining workflow
And integration test to test the pretraining script, the motivation
of the pretraining is to enhance the model's understanding of YAML
syntax and semantics.

In order to make the github CI integration test pass, I change the
pretraining dataset and size of the dataset, as well as using CPU to
train.

There are NoneType tokens when using the pretraining model for
inferencing, filter out those NoneType tokens before decoding.

It is possible that generated YAML examples are not complete. To make
the generated YAML examples complete, we can provide a more detailed
prompt to guide the model or adjust the `max_length` parameter to
allow the model to generate longer sequences or adjust the generation
parameters like `temperature`, `top_k`, `top_p` to control the randomness
and creativity of the output.

It is also possible encounter the TypeError decoding sequence, it is
related to the dataset and the training loss is not converged yet.

Signed-off-by: Wen Liang <liangwen12year@gmail.com>