Add user interface with ratings on the generated output
Use a Gradio-based web interface for interacting with a language model
and allowing users to rate the generated responses. It includes
functionalities to generate responses based on user input, and save
user ratings for the responses. Additionally, it maintains the history
of the conversation on the page.

In order to have free permanent hosting and GPU upgrades, we can run
`gradio deploy` from Terminal to deploy to huggingface Spaces
(https://huggingface.co/spaces).

I also added argparse to handle command-line arguments:
  - Required arguments:
    --model_path: Specifies the path to the model directory.
    --rating_result_path: Specifies the path to save the rating results.

Signed-off-by: Wen Liang <liangwen12year@gmail.com>