Explain the rating
The rating is within 1 to 5, with 5 being the highest rating value.
The rating stands for the quality of the generated response compared to
the expected response if the user is confident to vote. The user input,
generated result, and ratings will be stored in a rating file. And the
rating file will be used as the data for creating the reinforcement
learning reward model which will enable the language model to generate
the response has higher user ratings when similar user input was fed by
user.

Signed-off-by: Wen Liang <liangwen12year@gmail.com>